refactor: AccessTokenResolver 분리
