+++ tuk-api/src/main/kotlin/nexters/tuk/ui/resolver/Authenticated.kt
@@ -1,0 +1,5 @@
+package nexters.tuk.ui.resolver
+
+@Target(AnnotationTarget.VALUE_PARAMETER)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Authenticated
